Add apihelpers/UploadDataProviders.java to the API
Currently, `apihelpers/UploadDataProviders` exist as part of the API helper target on Chromium
as seen here (https://source.chromium.org/chromium/chromium/src/+/main:components/cronet/android/BUILD.gn;l=283). However,
we have already shipped `apihelpers/UploadDataProviders` as part of the API in AOSP so we have to keep shipping it. In order
to do that without introducing manually maintained Android.bp modules, add `apihelpers/UploadDataProviders` to gn2bp.

Ideally, in the future; we should ship both the API + API helpers but this is not yet decided. Once this is decided, this
change can be reverted

Bug: 314275954
Change-Id: I62df0892445d1073b38507d24125c20369780cae