Add cherry-picked CLs to gen_desc_json
Previously we had to cherry-pick those CLs manually in the desired branch which
introduced toil to the person performing the import, this should automate it and
we will be responsible for keeping this list up to date.

Although the desc_*.json files has changed, the Android.bp.swp did not change and
I confirmed that by running gen_android_bp.

Test: update_result.sh and make sure nothing changes.
Change-Id: I67c84f09a1241a0419fe34ec391cfc07cfcd5d65